Finished all pseudocode prep and create PECEncoder
All the necessary methods required for exporting are now pseudo coded
out, with the exception of several planned helper methods that will be
created as they are needed in the encoding process. Created the
PECEncoder class to house the logic for PEC export Encoding, the rest of
the functionality will exist in the PESFormat class.