Began PES export method construction.
Began constructing the PES export methods. This section in particular
deals with the PEC location and it's pointer.

*Had to do some research on writing unsigned values to output in Java.
Some of the data written to file needs to be unsigned. Java does not
have unsigned types, so bit masking will need to be used for unsigned
writes. I'll need to test thoroughly to ensure I do this correctly.