XML Adapter classes prepping for data binding
Attempted to get JAXB implemented for XML saving and loading. JavaFx
shapes can’t be directly bound in XML either, so I created a set of XML
adapter classes (XMLPatternAdapter, XMLStitchAdapter, XMLThreadAdapter,
and XMLShapeAdater) These classes collect and bind the necessary pattern
information for saving. The classes and their methods are done and build
clean. I will test this tomorrow and then hopefully move on to loading
the data back into a pattern.